Debugged Customer webservices

+++ webservice/src/main/java/com/jac/webservice/repository/CustomerRepository.java
@@ -4,10 +4,11 @@ import com.jac.webservice.exceptions.DatabaseException;
 import com.jac.webservice.exceptions.ItemExistException;
 import com.jac.webservice.exceptions.RecordDoesNotExistInDatabaseException;
 import com.jac.webservice.model.Customer;
+import org.apache.commons.lang3.exception.ExceptionUtils;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataRetrievalFailureException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
+
 import java.lang.reflect.Field;
 import java.sql.Date;
 import java.time.LocalDate;
@@ -22,45 +23,79 @@ public class CustomerRepository {
     @Autowired
     CustomerRowMapper customerRowMapper;
 
+    //    a helper function that checks if a customer exists by id
+    public Boolean checkIfCustomerExistByID(int id) {
+        String sql = "SELECT count(*) FROM customers WHERE id = ?";
+        int count = jdbcTemplate.queryForObject(sql, Integer.class, id);
+
+        if (count == 0) {
+            return false;
+        }
+        return true;
+    }
+
+    //    a helper function that checks for the number of customers by info in the database
+    public int countCustomerByInfo(String firstName, String lastName, LocalDate dob) {
+        String sql = "SELECT count(*) FROM customers WHERE first_name = ? AND last_name=? AND date_of_birth=?";
+        int count = jdbcTemplate.queryForObject(sql, Integer.class, firstName, lastName, Date.valueOf(dob));
+
+        return count;
+    }
+
+    //    a helper function that checks if a customer exists by Info in database
+    public Boolean checkIfCustomerExistByInfo(String firstName, String lastName, LocalDate dob) {
+        int count = countCustomerByInfo(firstName, lastName, dob);
+
+        if (count == 0) {
+            return false;
+        }
+        return true;
+    }
+
     //  A function that returns all customers in the database
     public List<Customer> getAllCustomers() {
-
-        try {
-            String sql = "SELECT * FROM customers";
-            return jdbcTemplate.query(sql, new CustomerRowMapper());
-
-        } catch (Exception exception) {
-            throw new DataRetrievalFailureException("An Exception occurred in CustomerRepository.getAllCustomers()");
-        }
+        String sql = "SELECT * FROM customers";
+        return jdbcTemplate.query(sql, new CustomerRowMapper());
     }
 
     //  A function that returns a customer from the database by its id
     public Customer getCustomerById(int id) {
-        try {
-            String sql = "SELECT * FROM customers WHERE id = ?";
-            return jdbcTemplate.queryForObject(sql, customerRowMapper, id);
-        } catch (Exception exc) {
-            throw new DataRetrievalFailureException("An Exception occurred in CustomerRepository.getBookById" + id);
+//        String sql = "SELECT count(*) FROM customers WHERE id = ?";
+//        int count = jdbcTemplate.queryForObject(sql, Integer.class, id);
+
+        if (!checkIfCustomerExistByID(id)) {
+            throw new RecordDoesNotExistInDatabaseException("Customer does not exist in the database!");
         }
+
+        String sql = "SELECT * FROM customers WHERE id = ?";
+        return jdbcTemplate.queryForObject(sql, customerRowMapper, id);
     }
 
     //  A function that returns a customer from the database by its name and DOB
     public Customer getCustomerByInfo(String firstName, String lastName, LocalDate dob) {
-        try {
-            String sql = "SELECT * FROM customers WHERE first_name = ? AND last_name=? AND date_of_birth=?";
-            return jdbcTemplate.queryForObject(sql, customerRowMapper, firstName, lastName, Date.valueOf(dob));
-        } catch (Exception exc) {
-            throw new DataRetrievalFailureException("An Exception occurred in CustomerRepository.getCustomerByInfo");
+
+        int count = countCustomerByInfo(firstName, lastName, dob);
+
+        if (count == 0) {
+            throw new RecordDoesNotExistInDatabaseException("Customer does not exist in the database!");
         }
+
+        if (count > 1) {
+            throw new DatabaseException(count + "customers with the same Info exist in the database!");
+        }
+
+        String sql = "SELECT * FROM customers WHERE first_name = ? AND last_name=? AND date_of_birth=?";
+        return jdbcTemplate.queryForObject(sql, customerRowMapper, firstName, lastName, Date.valueOf(dob));
     }
 
     //  A function that updates a customer in the database by its id
     public Customer updateCustomerById(int id, Customer customer) throws IllegalAccessException {
-        Customer fetchedCustomer = getCustomerById(id);
 
-        if(fetchedCustomer == null) {
-            throw new RecordDoesNotExistInDatabaseException("Failed to retrieve a Loan with id " + id);
+        if (!checkIfCustomerExistByID(id)) {
+            throw new RecordDoesNotExistInDatabaseException("Customer does not exist in the database!");
         }
+
+        Customer fetchedCustomer = getCustomerById(id);
 
         // returns the array of Field objects
         for (Field f : customer.getClass().getFields()) {
@@ -73,61 +108,75 @@ public class CustomerRepository {
         try {
             jdbcTemplate.update(
                     "UPDATE customers set " +
-                            "firs_name=?," +
-                            "middle_name=?" +
-                            "last_name=?"+
+                            "first_name=?," +
+                            "middle_name=?," +
+                            "last_name=?," +
                             "date_of_birth=?," +
-                            "address=? " +
-                            "city=? " +
-                            "state=? " +
-                            "zipcode=? " +
-                            "phone=? " +
+                            "address=?," +
+                            "city=?, " +
+                            "state=?, " +
+                            "zipcode=?, " +
+                            "phone=?, " +
                             "registration_date=? " +
                             "WHERE id=?",
-                    customer.getFirstName(), customer.getMiddleName(),customer.getLastName(),
-                    Date.valueOf(customer.getDob()), customer.getAddress(),customer.getCity(),
-                    customer.getState(),customer.getZipcode(),customer.getPhone(),
-                    Date.valueOf(customer.getRegistrationDate()),
+                    customer.getFirstName(), customer.getMiddleName(), customer.getLastName(),
+                    (customer.getDob() == null) ? customer.getDob() : Date.valueOf(customer.getDob()),
+                    customer.getAddress(), customer.getCity(),
+                    customer.getState(), customer.getZipcode(), customer.getPhone(),
+                    (customer.getRegistrationDate() == null) ? customer.getRegistrationDate() : Date.valueOf(customer.getRegistrationDate()),
                     id);
+
             return getCustomerById(id);
+
         } catch (Exception exc) {
             throw new DatabaseException(exc.getCause());
         }
     }
 
+
     //  A function that saves a new customer into the database
     public Customer saveCustomer(Customer customer) {
-        Customer fetchedCustomer = getCustomerByInfo(customer.getFirstName(), customer.getLastName(), customer.getDob());
-        if(fetchedCustomer != null) {
-            throw new ItemExistException("A customer with the same info already exists in the repository");
+
+        Boolean customerExists = checkIfCustomerExistByInfo(customer.getFirstName(), customer.getLastName(), customer.getDob());
+
+        if (customerExists) {
+            throw new ItemExistException("Customer with the same info already exist in the database!");
         }
+
         try {
             jdbcTemplate.update("Insert into customers " +
-                            "firs_name, middle_name, last_name, date_of_birth, address, city," +
-                            "state, zipcode, phone, registration_date " +
-                    "values (?, ?, ?, ?, ?, ?, ?, ?, ?, ?);",
+                            "(first_name, middle_name, last_name, date_of_birth, address, city," +
+                            "state, zipcode, phone, registration_date) " +
+                            "values (?, ?, ?, ?, ?, ?, ?, ?, ?, ?);",
                     customer.getFirstName(), customer.getMiddleName(), customer.getLastName(),
-                    Date.valueOf(customer.getDob()), customer.getAddress(), customer.getCity(),
+                    (customer.getDob() == null) ? customer.getDob() : Date.valueOf(customer.getDob()),
+                    customer.getAddress(), customer.getCity(),
                     customer.getState(), customer.getZipcode(), customer.getPhone(),
-                    Date.valueOf(customer.getRegistrationDate())
+                    (customer.getRegistrationDate() == null) ? customer.getRegistrationDate() : Date.valueOf(customer.getRegistrationDate())
             );
             int id = jdbcTemplate.queryForObject("select max(id) from customers", Integer.class);
             return getCustomerById(id);
 
-        }
-        catch (Exception exc) {
-            throw new NullPointerException("An exception occurred in CustomerRepository.saveCustomer");
+        } catch (Exception exc) {
+            throw new DatabaseException(ExceptionUtils.getRootCause(exc));
         }
     }
 
+
     //  A function that deletes a customer from the database by its id
     public void deleteCustomerById(int id) {
+
+        if (!checkIfCustomerExistByID(id)) {
+            throw new RecordDoesNotExistInDatabaseException("Record does not exist in the database!");
+        }
+
         try {
             jdbcTemplate.update("delete from customers where id=?", id);
-        }
-        catch (Exception exc) {
+        } catch (Exception exc) {
             throw new DatabaseException(exc.getCause());
         }
     }
 
 }
+
+
